\name{analyze}
\alias{analyze}
\title{Function to analyze DNA sequences for recombination blocks and estimate dates
using a binomial mass functuon.}
\usage{
  analyzedData <- analyze(dna, mutrate=10e-8,
    window.size=100, step.size=1, block.detect.length=5)
}
\arguments{
  \item{x}{A HybRIDS object generated by loading in a FASTA
  sequence file with \code{\link{dna.data.perpare()}}. If x
  is not of this class an error will be generated.}

  \item{mutrate}{A value representing the mutation rate you
  wish to specify, this is used in the analysis for the
  detection of potential hybrid blocks. By default this
  value is 10e-8.}

  \item{window.size}{A value representing the size of the
  sliding window, by default this is 100.}

  \item{step.size}{This value is the number of base pairs
  you want the sliding window yo jump along the sequences
  by. By default this is one.}

  \item{block.detect.length}{This value is the threshold
  number of consecutive homogenous snps between two
  sequences containing only informative sites, before
  HybRIDS considers the section of DNA as a potential
  region or block of recombination, and calculates it's P
  value and dating estimate. The default value is 5.}
}
\description{
  The function \code{analyze()} takes as it's primary
  argument the DNA data object loaded in with the function
  \code{\link{dna.data.prepare()}} and performs all
  analyses including distance mesausres along sliding
  windows, and initial detection and dating of potential
  blocks using the methods discussed in the main package
  documentation and publication. If none of the parameters
  are specified in the function call, then the default
  values shall be used. The output is a HybRIDS data object
  containing all nessecery information for subsequent
  plotting and anlysis.
}

