% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalAnalysis.R
\name{analyzeSignal}
\alias{analyzeSignal}
\title{Generic function used to perform a SignalAnalysis.}
\usage{
analyzeSignal(data, method)
}
\description{
Generic function used to perform a SignalAnalysis.
}
\section{Fields}{

\describe{
\item{\code{data}}{A variable containing the sequence data you wish to analyse.}

\item{\code{method}}{A variable which inherits from SignalAnalysisMethod. This
variable needs to contain all the options for the given SignalAnalysis method.}
}}

