% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SettingsClasses.R
\docType{class}
\name{BlockDatingSettings}
\alias{BlockDatingSettings}
\title{A reference class storing the settings for recombination block dating.}
\description{
A reference class storing the settings for recombination block dating.
}
\section{Fields}{

\describe{
\item{\code{MutationRate}}{Numeric vector of length one. Stores the mutation rate to be used when dating blocks.}

\item{\code{PValue}}{Numeric vector of length one. Stores the critical alpha value for testing the signifcance of recombination regions.}

\item{\code{BonfCorrection}}{Logical vector of length one, stores the option of whether the critical value stored in PValue will be corrected.}

\item{\code{DateAnyway}}{Logical vector of length one, sotres the option of whether blocks will be dated despite failing the critical alpha.}

\item{\code{MutationCorrection}}{Character vector of length 1, can be any of the model supported by the ape package. Default is "JC69".}
}}
\section{Methods}{

\describe{
\item{\code{setBonferonni(newBonf)}}{Set whether the critical value should be subject to bonferroni correction during block significance testing.}

\item{\code{setDateAnyway(newValue)}}{Set whether blocks should be kept and dated even if they fail the significance test.}

\item{\code{setMutationCorrection(model)}}{Set the model of sequence evolution to correct the distances/number of mutations used in block dating algorithm.}

\item{\code{setMutationRate(newRate)}}{Sets a new mutation rate for the settings.}

\item{\code{setPValue(newValue)}}{Set a new critical value for significance testing of recombinant blocks.}
}}

