% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BlockDetection.R
\docType{class}
\name{BlockDetectionSettings}
\alias{BlockDetectionSettings}
\title{Reference type for storing the settings for block detection.}
\description{
Reference type for storing the settings for block detection.
}
\section{Fields}{

\describe{
\item{\code{ManualThresholds}}{A numeric value between 1 and 100, the percentage raw sequence similarity a region has to reach, before it is identified as a block.}
}}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Initializes the settingsobject.}

\item{\code{setManualThresholds(values)}}{Checks input values for changing the manual sequence similarity thresholds for block detection and sets the parmeter.}

\item{\code{setSDstringency(value)}}{Checks input values for changing the SD stringency parameter for block detection and sets the parameter.}
}}

