\name{Calc.Similarity}
\alias{Calc.Similarity}
\title{Calc.Similarity}
\usage{
  Calc.Similarity(input, window.size = 100, step.size = 1)
}
\arguments{
  \item{input}{An object of type HybRIDSdna.}

  \item{window.size}{A number specifiying the size of the
  sliding window in base pairs. Must be integer.}

  \item{step.size}{A number specifying the size of the
  steps the sliding window takes, in base pairs. Must be
  integer.}
}
\description{
  Function performing the sequence similarity analysis with
  sliding window.
}

