% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TripletReference.R
\name{col_deter}
\alias{col_deter}
\title{Internal function to determine colours for the bars.}
\usage{
col_deter(invalues, reference)
}
\description{
Internal function to determine colours for the bars.
}

