% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ComparrisonSettings-class}
\alias{ComparrisonSettings}
\alias{ComparrisonSettings-class}
\title{ComparrisonSettings Reference Class.}
\description{
A Reference Class to represent a settings for Triplets.
}
\section{Fields}{

\describe{
\item{\code{Method}}{An integer vector of length 1.}
}}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Creates the object and sets all parameters to their default.}
}}

