% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ComparrisonSettings.}
\alias{ComparrisonSettings}
\alias{ComparrisonSettings.}
\title{A reference class to represent settings for triplet generation.}
\description{
A reference class to represent settings for triplet generation.
}
\section{Fields}{

\describe{
\item{\code{Method}}{An integer vector of length 1.}

\item{\code{DistanceThreshold}}{A numeric vector of length 1.}

\item{\code{PartitionStrictness}}{An integer vector of length 1.}

\item{\code{Refine}}{A logical vector of length 1.}

\item{\code{TripletCombinations}}{A list of integer vectors of length 3.}
}}
\section{Methods}{

\describe{
\item{\code{initialize()}}{Creates the object and sets all parameters to their default.}
}}

