% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{date.blocks}
\alias{date.blocks}
\title{Calculate the significance of identified recombinant blocks, and calculate their coalescence times.}
\usage{
date.blocks(blocksobj, dnaobj, mut, pthresh, bonfcorrect, danyway, model)
}
\arguments{
\item{blocksobj}{A data.frame containing all the information about the location of the identified block.}

\item{dnaobj}{A preprocessed DNAStringSet object.}

\item{mut}{A predefined mutation rate.}

\item{pthresh}{An alpha value threshold to use in significance testing.}

\item{bonfcorrect}{A logical value - whether or not to apply bonferroni correction to the pthresh.}

\item{danyway}{A logical value - if TRUE will esimate coalescence times of recombinant blocks even if they are not significant.}

\item{model}{A character value; denoting one of the mutational models allowed for in dist.dna of the APE package.}
}
\description{
Calculate the significance of identified recombinant blocks, and calculate their coalescence times.
}

