% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R
\name{excludeUnknownSites}
\alias{excludeUnknownSites}
\title{Subset sequence data objects such that they no longer contain N's}
\usage{
excludeUnknownSites(object)
}
\arguments{
\item{object}{An object which represents sequence data.}
}
\description{
Generic method to eliminate sites with 'N's from sequence data.
}

