% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTT.R
\name{fourTaxonTest}
\alias{fourTaxonTest}
\title{fourTaxonTest}
\usage{
fourTaxonTest(dna, fttRecord, numBlocks, lengthOfBlocks)
}
\description{
Computes the Four Taxon Test for a given set of P1, P2, P3, and P4.
Patterson's D is calculated, as is Fd from the paper Martin et al. (2014).
Fd is calculated for the scenario of complete introgression between P1 and P3,
and also for the scenario of complete introgression between P2 and P3.
The jack-knife implementation and computation of Z scores is based on that used in the software package
ANGSD.
}
