% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataStructures.R
\docType{class}
\name{FTTester}
\alias{FTTester}
\title{A Reference Class for performing and storing results of four taxon tests for introgression.}
\description{
A Reference Class for performing and storing results of four taxon tests for introgression.
}
\section{Fields}{

\describe{
\item{\code{global}}{Logical, whether a global statistic will be calculated for the four taxon tests.}

\item{\code{results}}{A list of reference objects defining the result of a given four taxon test.}
}}
\section{Methods}{

\describe{
\item{\code{getFTTs(selections)}}{Returns a list of references to FTTrecords objects according to user selection.}

\item{\code{initialize(dna)}}{Initialization method creates object with its default values.}
}}

