% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HybRIDSObject.R
\docType{class}
\name{HybRIDS}
\alias{HybRIDS}
\title{A Reference Class for managing a HybRIDS analysis.}
\description{
A Reference Class for managing a HybRIDS analysis.
}
\section{Fields}{

\describe{
\item{\code{DNA}}{A HybRIDSdna reference object.}
}}
\section{Methods}{

\describe{
\item{\code{initialize(dnafile = NULL)}}{Create HybRIDS object with default values for fields. The path to the FASTA file can be provided.}
}}

