% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Scan_Functions.R
\name{makeWindowFrames}
\alias{makeWindowFrames}
\title{Generate data-frame with co-ordinates of windows for analysis.}
\usage{
makeWindowFrames(winSize, stepSize, trackLen, bases)
}
\arguments{
\item{winSize}{Single numeric value. The size (in bp) of the sliding window.}

\item{stepSize}{Single numeric value. The number of base positions the
sliding window jumps along the sequences every time it moves.}

\item{trackLen}{Single numeric value. Length of the track the window will
travel down.}

\item{bases}{A numeric vector. The base positions that the sliding window
will slide over. Note in many HybridCheck analyses, the sliding window only
slides over the informative sites in a sequence alignment i.e. fully
conserved, non-informative sites are removed.}
}
\value{
slideFrame A data-frame of six columns. Each column is numeric.
}
\description{
Internal function. Used to make a data-frame of all the sliding
windows: their mid-points, start, ends. This data-frame is used during the
sequence alignment scans and also forms part of the final data table that
is produced.
}

