\name{plot.col.ref}
\alias{plot.col.ref}
\title{A function for plotting the reference colour palettes for the function
\code{\link{plot.all.contigs}}}
\usage{
  plot.col.ref(save=1)
}
\description{
  Only one argument to the function and that is the save
  argument. If set to 1 the plot generated will be saved to
  pdf file.
}

