\name{plot.HybRIDSdataRecord}
\alias{plot.HybRIDSdataRecord}
\title{Method for Plotting objects of classes associated with the HybRIDS package.}
\usage{
  \method{plot}{HybRIDSdataRecord} (x, linesplot = T,
    lineplot.legend = T, baseannotate = T, bpfreq = 500,
    mosaic.bars = T, mosaic.scale = 50, combine.plots = T,
    condense.mosaics = T, labfontsize = 14,
    legfontsize = 14)
}
\description{
  Method for Plotting objects of classes associated with
  the HybRIDS package.
}

