\name{plot.HybRIDSseqsim}
\alias{plot.HybRIDSseqsim}
\title{Method for Plotting objects of the class HybRIDSseqsim, associated with the HybRIDS package.}
\usage{
  \method{plot}{HybRIDSseqsim} (x, linesplot = T,
    legends = T, baseannotate = T, bpfreq = 500,
    densityplot = F, mosaic.bars = T, mosaic.scale = 50,
    condense.mosaics = T, labfontsize = 14,
    legfontsize = 14, onecanvas = T)
}
\description{
  Method for Plotting objects of the class HybRIDSseqsim,
  associated with the HybRIDS package.
}

