% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R, R/DNAStringSet_Methods.R
\docType{methods}
\name{polymorphicSites}
\alias{polymorphicSites}
\alias{polymorphicSites,DNAStringSet-method}
\title{Compute which sites in DNA sequence data are polymorphic.}
\usage{
polymorphicSites(object)

\S4method{polymorphicSites}{DNAStringSet}(object)
}
\arguments{
\item{object}{An object which represents sequence data.}
}
\description{
Generic method to get the indicies of polymorphic sites in DNA sequence data objects.
}

