% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalAnalysis.R
\docType{class}
\name{ReferenceWindowScan-class}
\alias{ReferenceWindowScan-class}
\title{A class for a Sliding Window based method of analysing recombination signal,
by comparing sequence similarity of many sequences, against one reference.}
\description{
This class, and others which inherit from it, represent methods of analysing
recombination signal which utilise sliding windows to copute their values.
For example, the class ReferenceWindowScan inherits from this class.
}
\section{Slots}{

\describe{
\item{\code{width}}{An integer specifying the width of the sliding window in the
analysis in base pairs.}

\item{\code{step}}{An integer specifying the number of base positions by which the 
sliding window in the analysis will move between each calculation.}
}}

