\name{seq.similarity}
\alias{seq.similarity}
\title{Internal Function to takes a trio of sequences (dnain), and calculate the sequence similarity across a sliding window.}
\usage{
  seq.similarity(dnain, triplet, win.size, s.size,
    fulllength, cutbp, verbose)
}
\description{
  Internal Function to takes a trio of sequences (dnain),
  and calculate the sequence similarity across a sliding
  window.
}

