% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataStructures.R
\docType{class}
\name{SequenceInformation}
\alias{SequenceInformation}
\title{Reference class to store the information from resolving the DNA sequence information in a triplet.}
\description{
Reference class to store the information from resolving the DNA sequence information in a triplet.
}

