% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R, R/DNAStringSet_Methods.R
\docType{methods}
\name{sequenceLength}
\alias{sequenceLength}
\alias{sequenceLength,DNAStringSet-method}
\title{Get the length of a sequence alignment}
\usage{
sequenceLength(object)

\S4method{sequenceLength}{DNAStringSet}(object)
}
\arguments{
\item{object}{An object which represents sequence data.}
}
\description{
Generic method to get the sequence length from an object.
}

