% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R, R/DNAStringSet_Methods.R
\docType{methods}
\name{sequenceNames}
\alias{sequenceNames}
\alias{sequenceNames,DNAStringSet-method}
\title{Get the sequence names from the object}
\usage{
sequenceNames(object)

\S4method{sequenceNames}{DNAStringSet}(object)
}
\arguments{
\item{object}{An object which represents sequence data.}
}
\description{
Generic method to get the sequence names from an object.
}

