% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalAnalysis.R
\docType{class}
\name{SignalAnalysisMethod-class}
\alias{SignalAnalysisMethod-class}
\title{A Virtual class to which all methods of computing a signal analysis should
belong.}
\description{
As a virtual class, has no fields, and should not be instantiated.
}

