% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataStructures.R
\docType{class}
\name{SimilarityScan}
\alias{SimilarityScan}
\title{Reference class to store the results from sequence similarity analyses, and block detection runs for a given triplet.}
\description{
Reference class to store the results from sequence similarity analyses, and block detection runs for a given triplet.
}
\section{Fields}{

\describe{
\item{\code{TableFile}}{A length 1 character vector storing the temporary filepath of the datatable of sequence similarity table.}

\item{\code{Table}}{Accessor function for the datatable of sequence similarity that is stored on temporary files.}

\item{\code{WindowSizeUsed}}{Single integer value, stores the size of the sliding window used for the scan.}

\item{\code{StepSizeUsed}}{Single integer value, stores the size of the sliding window used for the scan.}
}}
\section{Methods}{

\describe{
\item{\code{blankTable()}}{Method clears the SS analysis results table.}

\item{\code{finalize()}}{Called when the object is destroyed, makes sure to delete the file saved in the system's temporary directory.}

\item{\code{initialize(HCDir)}}{Method initializes the object, generates temporary filenames for the sequence similarity table.}

\item{\code{tableIsBlank()}}{Returns TRUE, if the SS analysis table is blank and no results are contained in it.}
}}

