% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R, R/DNAStringSet_Methods.R
\docType{methods}
\name{sitesWithKnown}
\alias{sitesWithKnown}
\alias{sitesWithKnown,DNAStringSet-method}
\title{Compute which sites in DNA sequence data don't have 'N's.}
\usage{
sitesWithKnown(object)

\S4method{sitesWithKnown}{DNAStringSet}(object)
}
\arguments{
\item{object}{An object which represents sequence data.}
}
\description{
Generic method to get the indicies of known sites in DNA sequence data objects.
}

