% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R, R/DNAStringSet_Methods.R
\docType{methods}
\name{sitesWithUnknown}
\alias{sitesWithUnknown}
\alias{sitesWithUnknown,DNAStringSet-method}
\title{Compute which sites in DNA sequence data have 'N's.}
\usage{
sitesWithUnknown(object)

\S4method{sitesWithUnknown}{DNAStringSet}(object)
}
\arguments{
\item{object}{An object which represents sequence data.}
}
\description{
Generic method to get the indicies of unknown sites in DNA sequence data objects.
}

