% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R, R/DNAStringSet_Methods.R
\docType{methods}
\name{slidingPoly}
\alias{slidingPoly}
\alias{slidingPoly,DNAStringSet,integer,integer-method}
\title{Calculate sequence identity across sliding windows.}
\usage{
slidingPoly(object, windowSize, stepSize)

\S4method{slidingPoly}{DNAStringSet,integer,integer}(object, windowSize,
  stepSize)
}
\arguments{
\item{object}{An object which represents sequence data.}
}
\description{
Move a sliding window across two aligned sequences,
calculating sequence similarity at each step.
}

