% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalAnalysis.R
\docType{class}
\name{SSToReferenceWindowScan-class}
\alias{SSToReferenceWindowScan-class}
\title{Sliding window scan of recombination signal between many suery sequences 
and one reference sequence.}
\description{
This class represents and analysis of recombination signal computed by 
calculating sequence similarity across sliding windows, between one
reference sequence, and many query sequences. As such this class
inherits from ReferenceWindowScan, which inherits from SlidingWindow.
}
\section{Slots}{

\describe{
\item{\code{settings}}{The ReferenceWindowScan object used to provide settings
for the performed analysis.}
}}

