% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R, R/DNAStringSet_Methods.R
\docType{methods}
\name{statesPerBase}
\alias{statesPerBase}
\alias{statesPerBase,DNAStringSet-method}
\title{Compute the number of states at each position in sequence data.}
\usage{
statesPerBase(object)

\S4method{statesPerBase}{DNAStringSet}(object)
}
\arguments{
\item{object}{An object which represents sequence data.}
}
\description{
Generic method to get the number of states at each position in objects containing sequence data.
}

