% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R, R/DNAStringSet_Methods.R
\docType{methods}
\name{subsetSites}
\alias{subsetSites}
\alias{subsetSites,DNAStringSet,integer-method}
\title{Generic method Create an DNA sequence object which only contains a subset of positions.}
\usage{
subsetSites(object, index)

\S4method{subsetSites}{DNAStringSet,integer}(object, index)
}
\arguments{
\item{object}{An object which represents sequence data.}

\item{index}{A numeric vector of sites to subset.}
}
\description{
Create an DNA sequence object which only contains the positions indicated by index.
}

