% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TripletIndexes}
\alias{TripletIndexes}
\title{TripletIndexes}
\usage{
TripletIndexes(SSSet, sequence1, sequence2 = "Individual",
  sequence3 = "Pair")
}
\arguments{
\item{SSSet}{An object that is either a HybRIDSseqsimSET, HybRIDSblockSET, or HybRIDSdatedBlocksSET}

\item{sequence1}{The name of the first sequence.}

\item{sequence2}{The name os the second sequence.}

\item{sequence3}{The same of the third sequence.}
}
\description{
A function to get the numeric index of the triplet containing the sequence names.
}
\details{
If a sequence name is provided for parameters sequence1, sequence2, and sequence3, then a single index corresponding to the triplet in which all three sequences occur is returned.
Otherwise if sequence1 and 2 are provided but sequence3 is not (so is left as "Pair"), several indicies are returned which are triplets containing the pair specified.
Furthuremore if sequence1 is provided (so sequence3 left as "Pair" and sequence 2 left as "Individual"),  several indicies are returned which are triplets containing the single seqence specified
}

