% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataStructures.R
\docType{class}
\name{Triplets}
\alias{Triplets}
\title{Reference class storing all triplets in a HC analysis.}
\description{
The Triplets reference class stores and manages operations over many Triplet objects.
}
\section{Methods}{

\describe{
\item{\code{deleteAllTriplets()}}{Removes all current triplet data completely.}

\item{\code{generateTriplets(dna, csettings, basefile)}}{Initializes all triplet objects based on the combination settings}

\item{\code{getAllIndexes()}}{Returns the indexes of the sequences (according to their rows in the HC sequence object) in each triplet as a list.}

\item{\code{getAllNames()}}{Returns the names of the sequences in each triplet as a list.}

\item{\code{getTriplets(selections)}}{Returns a list of references to triplets according to user selection.}

\item{\code{htmlSummary()}}{Prints a HTML summary of all the different triplet combinations.}

\item{\code{initialize(dna)}}{Initializes the object.}

\item{\code{tripletsGenerated()}}{Returns TRUE if the triplet generations have been prepared for.}
}}

