% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UserBlocks.R
\docType{class}
\name{UserBlocks}
\alias{UserBlocks}
\title{UserBlocks reference class}
\description{
The UserBlocks reference class is the class used to store and manipulate user defined blocks in HC.
}

