% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TripletReference.R
\name{vertbar_create}
\alias{vertbar_create}
\title{Internal function to create the vertical bars for mosaic plot.}
\usage{
vertbar_create(sequenceSimilarityTable, plottingFrameRow, whichComparrison)
}
\description{
Internal function to create the vertical bars for mosaic plot.
}

