% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Scan_Functions.R
\name{windowSizeChecker}
\alias{windowSizeChecker}
\title{Check sliding window settings used for a scan.}
\usage{
windowSizeChecker(winSize, trackLen)
}
\arguments{
\item{winSize}{A single numeric value. The proposed size of the sliding
window in the scan.}

\item{trackLen}{The A single numeric value. length of the track that the
window will travel down.}
}
\value{
winSize Single numeric value. The size of the sliding window to
be used in the analysis after the check. If an accepable size was chosen by
the user, this value will be identical to the input value.
}
\description{
Internal function. Used to make sure the user has entered
a sensible option for the sliding window size and step increments.
}
\examples{
HybridCheck:::windowSizeChecker(10, 100)
HybridCheck:::windowSizeChecker(100, 50)
HybridCheck:::windowSizeChecker(100, 10)
HybridCheck:::windowSizeChecker(100, 3)
}

