% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalAnalysis.R
\docType{methods}
\name{analyzeSignal}
\alias{analyzeSignal}
\alias{analyzeSignal,DNAMultipleAlignment,ReferenceWindowScan-method}
\alias{analyzeSignal,DNAStringSet,ReferenceWindowScan-method}
\title{Generic function used to perform a SignalAnalysis.}
\usage{
analyzeSignal(data, method)

\S4method{analyzeSignal}{DNAMultipleAlignment,ReferenceWindowScan}(data, method)

\S4method{analyzeSignal}{DNAStringSet,ReferenceWindowScan}(data, method)
}
\description{
Analyse the recombination signal present in a variable containing
sequence alignment data.
}
\details{
The specific method used to analyse the data is dispatched based on the
type of the input data, and the class of the method veriable.
}
\section{Methods (by class)}{
\itemize{
\item \code{data = DNAMultipleAlignment,method = ReferenceWindowScan}: Analyse recombination signal in a 
DNAMultipleAlignment, using the ReferenceWindowScan method.

\item \code{data = DNAStringSet,method = ReferenceWindowScan}: Analyse recombination signal in a 
DNAStringSet, using the ReferenceWindowScan method.
}}
\section{Fields}{

\describe{
\item{\code{data}}{A variable containing the sequence data you wish to analyse.}

\item{\code{method}}{A variable which inherits from SignalAnalysisMethod. This
variable needs to contain all the options for the given SignalAnalysis
method.}
}}

