\name{date.blocks}
\alias{date.blocks}
\title{Internal function that takes a dataframe of putative blocks between one sequence pair (e.g. A:B) for one threshold, detects significant blocks and dates them.}
\usage{
  date.blocks(blocksobj, dnaobj, mut, pair, pthresh)
}
\description{
  Internal function that takes a dataframe of putative
  blocks between one sequence pair (e.g. A:B) for one
  threshold, detects significant blocks and dates them.
}

