% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Methods.R, R/SignalAnalysis.R
\docType{methods}
\name{sequenceLength}
\alias{sequenceLength}
\alias{sequenceLength,SignalAnalysis-method}
\title{Get the length of the aligned sequences.}
\usage{
sequenceLength(object)

\S4method{sequenceLength}{SignalAnalysis}(object)
}
\arguments{
\item{object}{The object storing the multiple sequence alignment data from
which you wish to get the length.}
}
\value{
A numeric value representing the length of the aligned sequences.
All sequences in the alignment have the same length.
}
\description{
Get the length of the aligned sequences.
}
\section{Methods (by class)}{
\itemize{
\item \code{SignalAnalysis}: Get the length of the sequence alignment that was
analysed by the analysis represented by the given SignalAnalysis object.
}}

