% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\docType{methods}
\name{sequenceLength}
\alias{sequenceLength}
\title{Get the length of the aligned sequences.}
\usage{
sequenceLength(object)
}
\arguments{
\item{object}{The object storing the multiple sequence alignment data from which
you wish to get the length.}
}
\value{
A numeric value representing the length of the aligned sequences.
All sequences in the alignment have the same length.
}
\description{
Get the length of the aligned sequences.
}

