% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vertbar_create}
\alias{vertbar_create}
\title{Internal function to create the vertical bars for mosaic plot.}
\usage{
vertbar_create(sequenceSimilarityTable, plottingFrameRow, whichComparrison)
}
\description{
Internal function to create the vertical bars for mosaic plot.
}

